dynamic gates define with middleware

+++ app/Http/Controllers/Backend/RoleController.php
@@ -7,6 +7,7 @@ use App\Models\Role;
 use App\Models\Module;
 use Illuminate\Http\Request;
 use Illuminate\Support\Str;
+use Illuminate\Support\Facades\Gate;
 
 class RoleController extends Controller
 {
@@ -17,6 +18,7 @@ class RoleController extends Controller
      */
     public function index()
     {
+        Gate::authorize('app.roles.index');
         $roles = Role::all();
         return view('backend.roles.index', compact('roles'));
     }
@@ -28,6 +30,7 @@ class RoleController extends Controller
      */
     public function create()
     {
+        Gate::authorize('app.roles.create');
         $modules = Module::all();
         return view('backend.roles.form', compact('modules'));
     }
@@ -40,6 +43,7 @@ class RoleController extends Controller
      */
     public function store(Request $request)
     {
+        Gate::authorize('app.roles.create');
         $this->validate($request, [
             'name' => 'required|unique:roles',
             'permissions' => 'required|array',
@@ -48,7 +52,9 @@ class RoleController extends Controller
         Role::create([
             'name' => $request->name,
             'slug' => Str::slug($request->name, '-'),
-        ])->permissions()->sync($request->input('permissions'), []);
+        ])->permissions()->sync($request->permissions);
+        notify()->success("Role Added","Success");    
+
         return redirect()->route('app.roles.index');    
     }
 
@@ -71,6 +77,7 @@ class RoleController extends Controller
      */
     public function edit(Role $role)
     {
+        Gate::authorize('app.roles.edit');
         $modules = Module::all();
         return view('backend.roles.form', compact('modules', 'role'));
     }
@@ -84,11 +91,14 @@ class RoleController extends Controller
      */
     public function update(Request $request, Role $role)
     {
+        Gate::authorize('app.roles.edit');
         $role->update([
             'name' => $request->name,
             'slug' => Str::slug($request->name, '-')
         ]);
-        $role->permissions()->sync($request->input('permissions'), []);
+        $role->permissions()->sync($request->permissions);
+        notify()->success("Role Updated","Success");   
+
         return redirect()->route('app.roles.index');  
     }
 
@@ -100,10 +110,12 @@ class RoleController extends Controller
      */
     public function destroy(Role $role)
     {
+        Gate::authorize('app.roles.destroy');
         if($role->deletable){
             $role->delete();
+            notify()->success("Role Deleted","Success");  
         } else {
-
+            notify()->error("You can\'t delete system role","Error");
         }
         return back();
     }
